% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{density}
\alias{density}
\title{The internal functions for sparsenetgls}
\usage{
density(A, p)
}
\arguments{
\item{A}{A matrix for counting the edges in density() function}

\item{p}{Number of dimension in density() function}
}
\value{
Return half no of the links of Matrix A
}
\description{
Internal sparsenetgls functions
}
\details{
It is not to be called by users.
}
